requirements.txt
PyQt6>=6.8.0
python-dotenv>=1.0.0
mistralai==0.0.12
mutagen>=1.47.0
requests>=2.31.0 